+++ src/git_cache/git_command.py
@@ -97,7 +97,7 @@ def handle_git_command(called_as: List[str], args: List[str]) -> None:
         sys.exit(git_checkout(git_options))
 
     elif git_options.get_command() == "clone":
-        sys.exit(git_clone(git_options))
+        sys.exit(git_clone(called_as, git_options))
 
     elif git_options.get_command() == "lfs_fetch":
         sys.exit(git_lfs_fetch(git_options))